抄表模块静态测试修改完成

+++ src/main/java/com/yx/model/PropertyInfo.java
@@ -50,7 +50,7 @@ public class PropertyInfo implements Serializable {
 
     private Integer status;
 
-    private Integer houseId;
+    private Long houseId;
 
     private String remarks;
